requirements.txt
appnope==0.1.3
asgiref==3.5.0
asttokens==2.2.1
attrs==21.4.0
awsebcli==3.20.3
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
blessed==1.19.1
boto3==1.28.53
botocore==1.31.53
cached-property==1.5.2
cement==2.8.2
certifi==2021.10.8
cffi==1.15.0
chardet==5.1.0
charset-normalizer==2.0.12
click==8.1.5
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.3
comm==0.1.3
config==0.5.1
contourpy==1.1.0
cryptography==36.0.2
cssselect==1.2.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.6
Django==4.0.2
django-cors-headers==4.0.0
django-storages==1.14
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
ec2instanceconnectcli==1.0.2
et-xmlfile==1.1.0
executing==1.2.0
filelock==3.8.0
Fiona==1.9.4.post1
fonttools==4.41.0
future==0.16.0
geopandas==0.13.2
idna==3.3
ipykernel==6.24.0
ipython==8.14.0
jedi==0.18.2
jmespath==0.10.0
jsonschema==3.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
lxml==4.9.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.25.0
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
paramiko==2.10.3
parso==0.8.3
pathspec==0.9.0
pdfminer.six==20221105
pdfquery==0.4.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
pkgconfig==1.5.5
platformdirs==2.5.2
prompt-toolkit==3.0.39
psutil==5.9.5
psycopg2==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
PyJWT==2.8.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyproj==3.6.0
pyquery==2.0.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.5
pytz==2023.3
pyvips==2.2.1
PyYAML==5.4.1
pyzmq==25.1.0
requests==2.31.0
roman==4.1
s3transfer==0.6.2
semantic-version==2.8.5
shapely==2.0.1
six==1.14.0
soupsieve==2.4.1
sqlparse==0.4.2
stack-data==0.6.2
termcolor==1.1.0
texttable==1.6.4
tornado==6.3.2
traitlets==5.9.0
tzdata==2023.3
urllib3==1.26.9
virtualenv==20.16.5
wcwidth==0.1.9
websocket-client==0.59.0
